kbs: allow kbs to get reference values over gRPC
Let's allow the KBS to use the RVPS interface of the AS over gRPC so
that they can go through the KBS whether the AS is in built-in mode or
not.

This PR adds the RVPS endpoints to the attestation gRPC. In the past, we
have reused the RVPS proto in the AS, but there are some drawbacks to
using two different protos. Specifically, it makes the logic of the
connection pool in the KBS more complex.

Semantically, some people might prefer to have two protos (one for
attestation and one for reference values), but you could also say that
one proto is for the RVPS and the other is for the AS.

Signed-off-by: Tobin Feldman-Fitzthum <tobin@ibm.com>