kbs: add endpoints for configuring RVPS
Now that we expose the RVPS configuration to the KBS, add an interface
to the KBS that clients can use to configure the RVPS.

This is a privileged configuration interface that requires the client to
be authorized. In theory configuring the RVPS could be unprivileged, but
given that the sample extractor is always enabled, it is safest to do it
this way.

Signed-off-by: Tobin Feldman-Fitzthum <tobin@ibm.com>