policy: add a default gpu policy
Let's lay the groundwork for providing a policy to evaluate GPU
evidence. Today the EAR broker will expect to find a GPU policy called
`default_gpu` assuming it is evaluating the GPU Tee class. Currently,
the user must provide this policy themselves, but we can make things a
little easier by shipping a default.

Soon we can update this default with policies for various GPUs.

Signed-off-by: Tobin Feldman-Fitzthum <tobin@ibm.com>