verifier: allow verifier to return multiple sets of tee claims
The verifiers now return a vector of (tee_claims, tee_class).

This allows one verifier to represent multiple devices. These devices
are expected to share the same TeeType, which means that they need to
come from the same attester, but they could be different types of
devices (such as an nvidia gpu and an nvidia network switch).

In the future we could support more sophisticated mappings of attesters
to verifiers, but for now this is simple way to support more complex
devices.

Signed-off-by: Tobin Feldman-Fitzthum <tobin@ibm.com>